backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.9",
    "psycopg>=3.2.3",
    "pydantic-settings>=2.5.2",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0.35",
    "pyjwt>=2.8.0,<3.0.0",
    "pyjwt>=2.8.0,<3.0.0",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [".git", ".ruff_cache", ".venv", ".vscode", "__pypackages__", "venv"]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.8
target-version = "py38"

[tool.ruff.lint]
select = ["ARG", "FAST", "PLC0105", "C", "E", "W", "N", "UP", "A", "RUF"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
